==> Work/Modern Ways/Open Subtitles/Sentence/Product/1-Spark DataFrame Sentence Tokenize And Clean.ipynb <==
### Spark Sentence Detector

This notebook makes sentence tokenize with pyspark techniques.

#!pip install sparknlp

from pyspark.ml import PipelineModel
from sparknlp.annotator import *
from sparknlp.base import *
import sparknlp
from pyspark.ml import Pipeline
import pandas as pd
import re
import nltk
from nltk import word_tokenize
#pd.set_option("display.max_colwidth", 0)

from pyspark.sql import SparkSession

builder = SparkSession.builder\
        .appName("Spark NLP Licensed")\
        .master("local[*]")\
        .config("spark.driver.memory", "24G")\
        .config("spark.driver.maxResultSize", "2048GB")\
        .config("spark.serializer", "org.apache.spark.serializer.KryoSerializer")\
        .config("spark.kryoserializer.buffer.max", "2000M")\
        .config("spark.jars.packages", "com.johnsnowlabs.nlp:spark-nlp_2.12:3.3.2")

spark = builder.getOrCreate()

spark

#textFiles = spark.sparkContext.wholeTextFiles("oncology_notes/*")

#df = textFiles.toDF(schema=['path','text'])
#df.show()

# custom character alphabet 
tr = re.compile(r"[abcçdefgğhıijklmnoöprsştuüvyzqxw0123456789\.\?\{\}\[\]\(\),:;!'\"\$`]+", re.IGNORECASE|re.UNICODE) # Turkish filters='!"#$%&()*+,-./:;<=>?@[\\]^_`{|}~\t\n0123456789'
en = re.compile(r"[abcdefghıijklmnopqrstxuvwyz0123456789\.\?\{\}\[\]\(\),:;!'\"\$`]+", re.IGNORECASE|re.UNICODE) # English
nl = re.compile(r"[abcdefghıijklmnopqrstxuvwyzāăēĕīĭōŏūŭ0123456789\.\?\{\}\[\]\(\),:;!'\"\$`]+", re.IGNORECASE|re.UNICODE) # Dutch (Flemenk)
fr = re.compile(r"[abcçdefghıijklmnopqrstxuvwyzàâæèéêëîïôœùûüÿ0123456789\.\?\{\}\[\]\(\),:;!'\"\$`]+", re.IGNORECASE|re.UNICODE) # French
de = re.compile(r"[abcdefghıijklmnopqrstxuvwyzäöüßé0123456789\.\?\{\}\[\]\(\),:;!'\"\$`]+", re.IGNORECASE|re.UNICODE) # Deutch
es = re.compile(r"[abcdefghıijklmnopqrstxuvwyzñáéíóú0123456789\.\?\{\}\[\]\(\),:;!'\"\$`]+", re.IGNORECASE|re.UNICODE) # Spanish (¿¡)
pt = re.compile(r"[abcçdefghıijklmnopqrstxuvwyzàáâãéêíóôõú0123456789\.\?\{\}\[\]\(\),:;!'\"\$`]+", re.IGNORECASE|re.UNICODE) # Portuguese
it = re.compile(r"[abcdefghıijklmnopqrstxuvwyzàéèìòùî0123456789\.\?\{\}\[\]\(\),:;!'\"\$`]+", re.IGNORECASE|re.UNICODE) # Italian
ar = re.compile(r"[ٿصؼۤڳڲؿڎػڠجڿ٬ٸؽؒؓطۄڀۂؘؔتٚڛےٝڜؖڦ٫ډ۰زۇٖۀ،لۓ۱عٮێڔ۶ؚۧۜڤۏ٤إٞٷؗۖ؈ژۣؕؑٴأۻڸۺگاڴڹۯ؉ْڌ٣؍ي٧؟ـٟړۅؐڶُىڽېًۢؠضۚڄٛڏٱۦ٩س٦ڼڂٔۘ٠ښٌٍ۬ٳ۾ٲږذۋٵٜ٘ڞڅںٗهڣۿپڒۥۗڋیؙم؞ثۨ۹ٹڵڪظٶۭ١۔ڭەڨحٕ؎ٺڷٰ۪۫۲ڻڥۛڑڟټآڡغګ؊ّ٢ٯڧڮ؏ۮ؋ؤ٪ؾڗۼق۟دکوِڰڐۃ۽ہفر۷ڇ٭چڝ۴بۈٽڕۡھ۸٨ةٓڃئ؛ڬٙڙڢڱۊَۆۉځ۠ۍۑۙڊنءڈٻش٥ڍ؇۵ك۳خ\.\?\{\}\[\]\(\)\"\$\ا0123456789,':;\"`]+", re.IGNORECASE|re.UNICODE) # Arabic

folder_name = "Turkish"  # English, Turkish, German, French, Spanish, Arabic, Portuguese, Dutch, Italian
 
if folder_name == "English":
    text_name = "en"
    custom_alp = en
elif folder_name == "Turkish":
    text_name = "tr"
    custom_alp = tr
elif folder_name == "German":
    text_name = "de"
    custom_alp = de
elif folder_name == "French":
    text_name = "fr"
    custom_alp = fr
elif folder_name == "Spanish":
    text_name = "es"
    custom_alp = es
elif folder_name == "Arabic":
    text_name = "ar"
    custom_alp = ar
elif folder_name == "Portuguese":
    text_name = "pt"
    custom_alp = pt
elif folder_name == "Dutch":
    text_name = "nl"
    custom_alp = nl
elif folder_name == "Italian":
    text_name = "it"
    custom_alp = it
else:
    pass
#custom_alp = tr  # en, tr, de, fr, es, ar, pt, nl, it
print(f"{folder_name}\n{text_name}\n{custom_alp}") 

##### Part Num Loop

part_number = 1
#text_name = "tr"
#custom_alp = tr
sentence_tokenize_name = f"Spark_Sentence_Tokenize{part_number}"

df = spark.read.text(f'/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Turkish/Pyspark/{text_name}{part_number}.txt').toDF('text') # text data must be without Text Pre Process

df.select("text").head(1)

df.rdd.getNumPartitions()

df = df.repartition(200)

df.rdd.getNumPartitions()

documenter = DocumentAssembler()\
    .setInputCol("text")\
    .setOutputCol("document")
    
sentencerDL = SentenceDetectorDLModel\
    .pretrained("sentence_detector_dl", "xx") \
    .setInputCols(["document"]) \
    .setOutputCol("sentences")

model = PipelineModel(stages=[documenter, sentencerDL])

result = model.transform(df)

result.show()

result.printSchema()

pd_res = result.select("sentences.result").toPandas()
pd_res

pd_res.to_csv(f"{sentence_tokenize_name}.csv", index=False)

#### Clean DataFrame Sentence

#!pip install modin[ray]

import ray  # ayrı olarak çalıştırılmalı
ray.init() 

import modin.pandas as pd
import unicodedata

df2 = pd.read_csv(f"{sentence_tokenize_name}.csv")
df2

def convert_one_character_letter(text):
    '''
    This function converts two byte occupy of letter to one byte unicode character without any visual change \n
    like as Turkish character ç,ş,ö,ğ. \n
    convert_one_character_letter(text): text is any string word or sentence.
    '''
    new_text = unicodedata.normalize('NFC', f"{text}")
    
    return new_text

#def clean_text(text):
#    apostro = re.compile(re.escape("\'"), re.IGNORECASE)
#    text2 = apostro.sub(" \'", text)
#    text3 = re.findall(custom_alp, text2)
#    text4 = " ".join(text3)
#    clean_brackets = text4.strip("[]")
#    clean_brackets_data = re.sub('[\(\[\{].*?[\)\]\}]', '', clean_brackets, re.UNICODE)
#    clean_punc = re.sub(r'[^\w\s]','',clean_brackets_data, re.UNICODE)
#    text_clean_list = re.findall(r'\w+', clean_punc, re.UNICODE)
#    #alpha_num = [w for w in text_clean_list if w.isalnum()]
#    text_result = (" ".join(text_clean_list)).lower()
#    
#    return text_result 

def clean_text(text):
    '''
    clean_text(text) text is plain text or sentence. This function uses custom alp character and specific tecniques for cleaning.\n
    text = "&ali av9 yapar%"\n
    clean_text(text)  
    '''
    apostro = re.compile(re.escape("\'"), re.IGNORECASE)
    text2 = apostro.sub(" \'", text)
    text_one = convert_one_character_letter(text2)
    text3 = re.findall(custom_alp, text_one)
    text4 = " ".join(text3)
    clean_brackets = text4.strip("[]")
    clean_brackets_data = re.sub('[\(\[\{].*?[\)\]\}]', '', clean_brackets, re.UNICODE)
    clean_punc = re.sub(r'[^\w\s]','',clean_brackets_data, re.UNICODE)
    text_clean_list = re.findall(r'\w+', clean_punc, re.UNICODE)
    #alpha_num = [w for w in text_clean_list if w.isalnum()]
    text_result = (" ".join(text_clean_list)).lower()

    return text_result 

#def clean_text(text):
#    clean_brackets = text.strip("[]")
#    clean_brackets_data = re.sub('[\(\[\{].*?[\)\]\}]', '', clean_brackets, re.UNICODE)
#    clean_punc = re.sub(r'[^\w\s]','',clean_brackets_data, re.UNICODE)
#    text_clean_list = re.findall(r'\w+', clean_punc, re.UNICODE)
#    #alpha_num = [w for w in text_clean_list if w.isalnum()]
#    text_result = (" ".join(text_clean_list)).lower()
#    return text_result 

sent_list = []
for x in df2.result:
    x = clean_text(x)
    sent_list.append(x)  

df_clean = pd.DataFrame(sent_list)
df_clean.head()

df_clean = df_clean.rename(columns={0:"sentence"})
df_clean

df_clean.drop(df_clean[df_clean.sentence == ""].index, inplace=True)

sent_count = df_clean.sentence.value_counts().sort_values(ascending=False)
#sent_count

sent_count_df = pd.DataFrame(sent_count).reset_index()
sent_count_df.rename(columns={"index":"sentence", "sentence":"frequency"}, inplace=True)

total_frequency = sent_count_df.frequency.sum()

sent_count_df["ratio"] = (sent_count_df.frequency/total_frequency)*100

def sentence_lenght(sentence):
    word_var2 = re.findall(r'\w+', sentence, re.UNICODE)
    if len(word_var2) <= 10:
        return sentence
    else:
        return "sentence is bigger than ten word"

sent_count_df.sentence = sent_count_df.sentence.apply(sentence_lenght)

sent_count_df.drop(sent_count_df[sent_count_df.sentence == "sentence is bigger than ten word"].index, axis=0, inplace=True)
sent_count_df.reset_index(drop=True, inplace=True)

df_result = sent_count_df[sent_count_df.frequency>1]

df_result.to_csv(f"Clean_{sentence_tokenize_name}.csv", index=False)
