==> Work/Modern Ways/Open Subtitles/Sentence/Product/3-Merge Word Sentence N Gram Rev1.ipynb <==
### Merge Word Sentence And N Gram

This notebook merges sentence tokenize files to one merge file.

import pandas as pd
import numpy as np
import re
import glob
import nltk
from nltk import word_tokenize

#### Cancat Axis 1

df1 = pd.read_csv("Clean_Spark_Sentence_Tokenize1.csv")
df1.head()

frequency_sum1 = df1.frequency.sum()
frequency_sum1

cumul_ratio_last1 = df1.iloc[-1].cumul_ratio
cumul_ratio_last1

total_frequency1 = ((100*frequency_sum1))/cumul_ratio_last1
total_frequency1

df1 = df1[["sentence","frequency"]]

df1 = df1.rename(columns={"frequency":"frequency1"})

df1 = df1.drop_duplicates(subset=['sentence'])
df1.reset_index(drop=True, inplace=True)

df2 = pd.read_csv("Clean_Spark_Sentence_Tokenize2.csv")
df2.head()

frequency_sum2 = df2.frequency.sum()
frequency_sum2

cumul_ratio_last2 = df2.iloc[-1].cumul_ratio
cumul_ratio_last2

total_frequency2 = ((100*frequency_sum2))/cumul_ratio_last2
total_frequency2

df2 = df2[["sentence","frequency"]]

df2 = df2.rename(columns={"frequency":"frequency2"})

df2 = df2.drop_duplicates(subset=['sentence'])
df2.reset_index(drop=True, inplace=True)

df_12 = pd.merge(df1, df2, how ='inner', on =['sentence', 'sentence'])
df_12

df_12.reset_index(drop=True, inplace=True)

df3 = pd.read_csv("Clean_Spark_Sentence_Tokenize3.csv")
df3.head()

frequency_sum3 = df3.frequency.sum()
frequency_sum3

cumul_ratio_last3 = df3.iloc[-1].cumul_ratio
cumul_ratio_last3

total_frequency3 = ((100*frequency_sum3))/cumul_ratio_last3
total_frequency3

df3 = df3[["sentence","frequency"]]

df3 = df3.rename(columns={"frequency":"frequency3"})

df3 = df3.drop_duplicates(subset=['sentence'])
df3.reset_index(drop=True, inplace=True)

df_123 = pd.merge(df_12, df3, how="inner", on="sentence")
df_123

df_123.reset_index(drop=True, inplace=True)

df4 = pd.read_csv("Clean_Spark_Sentence_Tokenize4.csv")
df4.head()

frequency_sum4 = df4.frequency.sum()
frequency_sum4

cumul_ratio_last4 = df4.iloc[-1].cumul_ratio
cumul_ratio_last4

total_frequency4 = ((100*frequency_sum4))/cumul_ratio_last4
total_frequency4

df4 = df4[["sentence","frequency"]]

df4 = df4.rename(columns={"frequency":"frequency4"})

df4 = df4.drop_duplicates(subset=['sentence'])
df4.reset_index(drop=True, inplace=True)

df_1234 = pd.merge(df_123, df4,how="inner", on="sentence")
df_1234

df_1234.reset_index(drop=True, inplace=True)

df5 = pd.read_csv("Clean_Spark_Sentence_Tokenize5.csv")
df5.head()

frequency_sum5 = df5.frequency.sum()
frequency_sum5

cumul_ratio_last5 = df5.iloc[-1].cumul_ratio
cumul_ratio_last5

total_frequency5 = ((100*frequency_sum5))/cumul_ratio_last5
total_frequency5

df5 = df5[["sentence","frequency"]]

df5 = df5.rename(columns={"frequency":"frequency5"})

df5 = df5.drop_duplicates(subset=['sentence'])
df5.reset_index(drop=True, inplace=True)

df_12345= pd.merge(df_1234, df5, how="inner", on="sentence")
df_12345

df_12345.reset_index(drop=True, inplace=True)

df6 = pd.read_csv("Clean_Spark_Sentence_Tokenize6.csv")
df6.head()

frequency_sum6 = df6.frequency.sum()
frequency_sum6

cumul_ratio_last6 = df6.iloc[-1].cumul_ratio
cumul_ratio_last6

total_frequency6 = ((100*frequency_sum6))/cumul_ratio_last6
total_frequency6

df6 = df6[["sentence","frequency"]]

df6 = df6.rename(columns={"frequency":"frequency6"})

df6 = df6.drop_duplicates(subset=['sentence'])
df6.reset_index(drop=True, inplace=True)

df_123456= pd.merge(df_12345, df6, how="inner", on="sentence")
df_123456

df_123456.reset_index(drop=True, inplace=True)

df7 = pd.read_csv("Clean_Spark_Sentence_Tokenize7.csv")
df7.head()

frequency_sum7 = df7.frequency.sum()
frequency_sum7

cumul_ratio_last7 = df7.iloc[-1].cumul_ratio
cumul_ratio_last7

total_frequency7 = ((100*frequency_sum7))/cumul_ratio_last7
total_frequency7

df7 = df7[["sentence","frequency"]]

df7 = df7.rename(columns={"frequency":"frequency7"})

df7 = df7.drop_duplicates(subset=['sentence'])
df7.reset_index(drop=True, inplace=True)

df_1234567= pd.merge(df_123456, df7, how="inner", on="sentence")
df_1234567

df_1234567 = df_1234567.drop_duplicates(subset=['sentence'])
df_1234567.reset_index(drop=True, inplace=True)

df_1234567["frequency"] = df_1234567.frequency1 + df_1234567.frequency2 + df_1234567.frequency3 + df_1234567.frequency4 + df_1234567.frequency5 + df_1234567.frequency6 + df_1234567.frequency7
df_1234567.drop(["frequency1","frequency2","frequency3","frequency4","frequency5","frequency6","frequency7"], axis=1, inplace=True)
df_1234567

df_1234567 = df_1234567.sort_values(by=['frequency'],ascending=False)
df_1234567

df_1234567.iloc[26]

df_1234567.to_excel("Spark_Turkish_Sentence_Result_Merge.xlsx", sheet_name='Sentence', index=False)  

total_frequency = total_frequency1 + total_frequency2 + total_frequency3 + total_frequency4 + total_frequency5 + total_frequency6 + total_frequency7
total_frequency

#### Concat Axis 0

df1 = pd.read_csv("/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Pyspark/Result/Clean_Spark_Sentence_Tokenize1.csv")
df1

df2 = pd.read_csv("/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Pyspark/Result/Clean_Spark_Sentence_Tokenize2.csv")
df2

df3 = pd.read_csv("/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Pyspark/Result/Clean_Spark_Sentence_Tokenize3.csv")
df3

df4 = pd.read_csv("/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Pyspark/Result/Clean_Spark_Sentence_Tokenize4.csv")
df4

df5 = pd.read_csv("/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Pyspark/Result/Clean_Spark_Sentence_Tokenize5.csv")
df5

df6 = pd.read_csv("/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Pyspark/Result/Clean_Spark_Sentence_Tokenize6.csv")
df6

df7 = pd.read_csv("/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Pyspark/Result/Clean_Spark_Sentence_Tokenize7.csv")
df7

tokenize_list = glob.glob("/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Pyspark/Result/Sentence/Sentence Tokenize Data/Clean_Spark_Sentence_Tokenize*.csv")
tokenize_list

# Read all data
data_kind = "sentence"
df_all_result = pd.DataFrame()
for i in tokenize_list:
#for i in range(1,8):
    df_var = pd.read_csv(f'{i}')
    #df_var = pd.read_csv(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Pyspark/Result/Clean_Spark_Sentence_Tokenize{i}.csv")
    df_all_result = pd.concat([df_var, df_all_result], axis=0)
    df_all_result.dropna(inplace=True)
    df_all_result.reset_index(drop=True, inplace=True)
df_all_result = df_all_result[[f"{data_kind}", "frequency"]]
df_all_result

df_all_group = df_all_result.groupby([f"{data_kind}"])["frequency"].sum().sort_values(ascending=False)
df_all_group = df_all_group.reset_index()
df_all_group.dropna(inplace=True)
df_all_group.reset_index(drop=True, inplace=True)
df_all_group

df_all_group = df_all_group[df_all_group.frequency > 1]

#total_frequency = df_all_group.frequency.sum()

#df_all_group["ratio"] = (df_all_group.frequency/total_frequency)*100

#df_all_group["cumul_ratio"] = np.cumsum(df_all_group["ratio"])

df_all_group.to_csv(f"{data_kind.capitalize()}_Merge.csv", index=False)
