==> Work/Modern Ways/Open Subtitles/Sentence/Product/5-Sentence DataFrame Sentence Word Length Grouping Rev2.ipynb <==
### Sentence Data Frame Sentence Word Length Grouping

This notebook, which creates sentence groups according to sentence word length, is used to extract ngrams and ngram-sentences. 

import os
import multiprocessing
#import multiprocessing as mp
from multiprocessing import Process, Manager, Pool, Queue
import pandas as pd
import numpy as np
import nltk
from nltk import word_tokenize

nprocs = multiprocessing.cpu_count()
print(f"Number of CPU cores: {nprocs}")

#### Sentence More Than Variable Word

sent_path = "/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Pyspark/Result/Clean Spark Sentence Merge/Clean_Spark_Sentence_Merge_Result.csv"

df_all_sent = pd.read_csv(f"{sent_path}")
df_all_sent = df_all_sent[["sentence","frequency"]]
df_all_sent

sent_list = df_all_sent.iloc[:,0].to_list()

# import os
# import multiprocessing
# from multiprocessing import Process, Manager, Pool, Queue
manager = multiprocessing.Manager()
sent1_list = manager.list()
sent2_list = manager.list()
sent3_list = manager.list()
sent4_list = manager.list()
sent5_list = manager.list()
sent6_list = manager.list()
sent7_list = manager.list()
sent8_list = manager.list()
sent9_list = manager.list()

def sent_more_than(sent):
    word = word_tokenize(sent)
    if len(word) > 9:
        sent9_list.append(sent)
        sent8_list.append(sent)
        sent7_list.append(sent)
        sent6_list.append(sent)
        sent5_list.append(sent)
        sent4_list.append(sent)
        sent3_list.append(sent)
        sent2_list.append(sent)
        sent1_list.append(sent)     
    elif len(word) > 8:
        sent8_list.append(sent)
        sent7_list.append(sent)
        sent6_list.append(sent)
        sent5_list.append(sent)
        sent4_list.append(sent)
        sent3_list.append(sent)
        sent2_list.append(sent)
        sent1_list.append(sent)
    elif len(word) > 7:
        sent7_list.append(sent)
        sent6_list.append(sent)
        sent5_list.append(sent)
        sent4_list.append(sent)
        sent3_list.append(sent)
        sent2_list.append(sent)
        sent1_list.append(sent)
    elif len(word) > 6:
        sent6_list.append(sent)
        sent5_list.append(sent)
        sent4_list.append(sent)
        sent3_list.append(sent)
        sent2_list.append(sent)
        sent1_list.append(sent)
    elif len(word) > 5:
        sent5_list.append(sent)
        sent4_list.append(sent)
        sent3_list.append(sent)
        sent2_list.append(sent)
        sent1_list.append(sent)
    elif len(word) > 4:
        sent4_list.append(sent)
        sent3_list.append(sent)
        sent2_list.append(sent)
        sent1_list.append(sent)
    elif len(word) > 3:
        sent3_list.append(sent)
        sent2_list.append(sent)
        sent1_list.append(sent)
    elif len(word) > 2:
        sent2_list.append(sent)
        sent1_list.append(sent)
    elif len(word) > 1:
        sent1_list.append(sent)
    else:
        pass
    
if __name__ == '__main__':
    # with Pool(16) as p:
    with Pool(nprocs) as p: # Pool number CPU sayısına eşit olursa tüm CPU lar çalışır
        p.map(sent_more_than, sent_list)

df_sent1 = pd.DataFrame(list(sent1_list))
df_sent1.rename(columns={0:"sentence"}, inplace=True)

df_more_one_word = pd.merge(df_all_sent, df_sent1, how='inner', on="sentence")
df_more_one_word

df_more_one_word.to_csv("Sentence_More_Than_One_Word.csv", index=False)

df_sent2 = pd.DataFrame(list(sent2_list))
df_sent2.rename(columns={0:"sentence"}, inplace=True)

df_more_two_word = pd.merge(df_all_sent, df_sent2, how='inner', on="sentence")
df_more_two_word

df_more_two_word.to_csv("Sentence_More_Than_Two_Word.csv", index=False)

df_sent3 = pd.DataFrame(list(sent3_list))
df_sent3.rename(columns={0:"sentence"}, inplace=True)

df_more_three_word = pd.merge(df_all_sent, df_sent3, how='inner', on="sentence")
df_more_three_word

df_more_three_word.to_csv("Sentence_More_Than_Three_Word.csv", index=False)

df_sent4 = pd.DataFrame(list(sent4_list))
df_sent4.rename(columns={0:"sentence"}, inplace=True)

df_more_four_word = pd.merge(df_all_sent, df_sent4, how='inner', on="sentence")
df_more_four_word

df_more_four_word.to_csv("Sentence_More_Than_Four_Word.csv", index=False)

df_sent5 = pd.DataFrame(list(sent5_list))
df_sent5.rename(columns={0:"sentence"}, inplace=True)

df_more_five_word = pd.merge(df_all_sent, df_sent5, how='inner', on="sentence")
df_more_five_word

df_more_five_word.to_csv("Sentence_More_Than_Five_Word.csv", index=False)

df_sent6 = pd.DataFrame(list(sent6_list))
df_sent6.rename(columns={0:"sentence"}, inplace=True)

df_more_six_word = pd.merge(df_all_sent, df_sent6, how='inner', on="sentence")
df_more_six_word

df_more_six_word.to_csv("Sentence_More_Than_Six_Word.csv", index=False)

df_sent7 = pd.DataFrame(list(sent7_list))
df_sent7.rename(columns={0:"sentence"}, inplace=True)

df_more_seven_word = pd.merge(df_all_sent, df_sent7, how='inner', on="sentence")
df_more_seven_word

df_more_seven_word.to_csv("Sentence_More_Than_Seven_Word.csv", index=False)

df_sent8 = pd.DataFrame(list(sent8_list))
df_sent8.rename(columns={0:"sentence"}, inplace=True)

df_more_eight_word = pd.merge(df_all_sent, df_sent8, how='inner', on="sentence")
df_more_eight_word

df_more_eight_word.to_csv("Sentence_More_Than_Eight_Word.csv", index=False)

df_sent9 = pd.DataFrame(list(sent9_list))
df_sent9.rename(columns={0:"sentence"}, inplace=True)

df_more_nine_word = pd.merge(df_all_sent, df_sent9, how='inner', on="sentence")
df_more_nine_word

df_more_nine_word.to_csv("Sentence_More_Than_Nine_Word.csv", index=False)
