==> Work/Modern Ways/Open Subtitles/Sentence/Product/5-Two DataFrame Compare Columns Difference Rev1.ipynb <==
### Compare DataFrame Columns

This notebook is used for exract variable sentences.

import pandas as pd
import numpy as np

#### One Gram

sent_path = "/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Pyspark/Result/Clean Spark Sentence Merge/Clean_Spark_Sentence_Merge_Result.csv"

df_all_sent = pd.read_csv(f"{sent_path}")
df_all_sent = df_all_sent[["sentence","frequency"]]
df_all_sent

df_one_gram = pd.read_csv("One_Gram_And_Sentence.csv") # one_gram and sentence
df_one_gram  

#df_all_sent["match"] = np.where(df_all_sent.sentence == df_one_gram.sentence, True, False)

set_all_sent = set(df_all_sent.iloc[:,0])

set_one_gram = set(df_one_gram.iloc[:,0])

set_more_one = set_all_sent.difference(set_one_gram)

df_more_one = pd.DataFrame(set_more_one)
df_more_one.rename(columns={0:"sentence"}, inplace=True)
df_more_one

df_more_one_word = pd.merge(df_all_sent, df_more_one, how='inner', on="sentence")
df_more_one_word

df_more_one_word.to_csv("Sentence_More_Than_One_Word.csv", index=False)

#### Two Gram

df_more_one_word = pd.read_csv("Sentence_More_Than_One_Word.csv")
df_more_one_word

df_two_gram = pd.read_csv("/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Pyspark/Data/Sentence/Two_Gram_And_Sentence.csv") # two_gram and sentence
df_two_gram

set_more_one_word = set(df_more_one_word.iloc[:,0])

set_two_gram = set(df_two_gram.iloc[:,0])

set_more_two = set_more_one_word.difference(set_two_gram)

df_more_two = pd.DataFrame(set_more_two)
df_more_two.rename(columns={0:"sentence"}, inplace=True)
df_more_two

df_more_two_word = pd.merge(df_more_one_word, df_more_two, how='inner', on="sentence")
df_more_two_word

df_more_two_word.to_csv("Sentence_More_Than_Two_Word.csv", index=False)

#### Three Gram

df_more_two_word = pd.read_csv("Sentence_More_Than_Two_Word.csv")
df_more_two_word

df_three_gram = pd.read_csv("/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Pyspark/Data/Sentence/Three_Gram_And_Sentence.csv")
df_three_gram

set_more_two_word = set(df_more_two_word.iloc[:,0])
set_three_gram = set(df_three_gram.iloc[:,0])
set_more_three = set_more_two_word.difference(set_three_gram)

df_more_three = pd.DataFrame(set_more_three)
df_more_three.rename(columns={0:"sentence"}, inplace=True)
df_more_three

df_more_three_word = pd.merge(df_more_two_word, df_more_three, how='inner', on="sentence")
df_more_three_word

df_more_three_word.to_csv("Sentence_More_Than_Three_Word.csv", index=False)

#### Four Gram

df_more_three_word = pd.read_csv("Sentence_More_Than_Three_Word.csv")
df_more_three_word

df_four_gram = pd.read_csv("/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Pyspark/Data/Sentence/Four_Gram_And_Sentence.csv")
df_four_gram

set_more_three_word = set(df_more_three_word.iloc[:,0])
set_four_gram = set(df_four_gram.iloc[:,0])
set_more_four = set_more_three_word.difference(set_four_gram)

df_more_four = pd.DataFrame(set_more_four)
df_more_four.rename(columns={0:"sentence"}, inplace=True)
df_more_four

df_more_four_word = pd.merge(df_more_three_word, df_more_four, how='inner', on="sentence")
df_more_four_word

df_more_four_word.to_csv("Sentence_More_Than_Four_Word.csv", index=False)

#### Five Gram

df_more_four_word = pd.read_csv("Sentence_More_Than_Four_Word.csv")
df_more_four_word

df_five_gram = pd.read_csv("/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Pyspark/Data/Sentence/Five_Gram_And_Sentence.csv")
df_five_gram

set_more_four_word = set(df_more_four_word.iloc[:,0])
set_five_gram = set(df_five_gram.iloc[:,0])
set_more_five = set_more_four_word.difference(set_five_gram)

df_more_five = pd.DataFrame(set_more_five)
df_more_five.rename(columns={0:"sentence"}, inplace=True)
df_more_five

df_more_five_word = pd.merge(df_more_four_word, df_more_five, how='inner', on="sentence")
df_more_five_word

df_more_five_word.to_csv("Sentence_More_Than_Five_Word.csv", index=False)
