==> Work/Modern Ways/Open Subtitles/Sentence/Product/6-Sentence Unique Code.ipynb <==
### Sentence Unique Code

This notebook is used for give unique code to each sentences.

import pandas as pd
import numpy as np
import nltk
from nltk import word_tokenize
import re

df_sent = pd.read_csv("/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Pyspark/Result/Sentence/Spark Turkish Sentence Merge/Clean_Spark_Sentence_Merge_Result.csv")
df_sent = df_sent.head(500)
df_sent

# # 1-Way
# count_index = 0
# code_list = []
# for i in df_sent.sentence:
#     g = len(re.findall("\w+", i))
#     l = len(i)
#     f = df_sent.iloc[count_index].frequency
#     unique_code = f"G{g}-L{l}-F{f}-{count_index}"
#     code_list.append(unique_code)
#     count_index += 1
# df_sent["code"] = pd.DataFrame(code_list)
# df_sent

# 2-Way
# df_empty = pd.DataFrame({'col_name' : []})
df_sent["code"] = pd.Series(dtype="string")
count_index = 0
for i in df_sent.sentence:
    g = len(re.findall("\w+", i))
    l = len(i)
    f = df_sent.iloc[count_index].frequency
    unique_code = f"G{g}-L{l}-F{f}-{count_index}"
    df_sent.at[count_index, "code"] = unique_code
    count_index += 1
df_sent

df_sent.iloc[33]

df_sent.to_csv("Clean_Spark_Sentence_Merge_Result_Code.csv", index=False)

#### Multi Processing Unique Code

import pandas as pd
import numpy as np
import os
import multiprocessing
#import multiprocessing as mp
from multiprocessing import Process, Manager, Pool, Queue
from itertools import islice
import re

help(multiprocessing)

nprocs = multiprocessing.cpu_count()
print(f"Number of CPU cores: {nprocs}")

file_name = "Sentence_Merge.csv"
file_path = "/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Turkish/Pyspark/Result/Sentence"

df_sent = pd.read_csv(f"{file_path}/{file_name}")
df_sent = df_sent.head(10000)
df_sent

index_sent = zip(df_sent.index, df_sent.iloc[:,0])
zip_list = list(index_sent) # index numarası ve sentence içeriyor
#zip_list

# index, sentence = zip(*index_sent)

sent_code_list = []

# import os
# from multiprocessing import Process, Manager, Pool, Queue
manager = multiprocessing.Manager()
sent_code_list = manager.list()


def sentence_code(index_num, sent):
    g = len(re.findall("\w+", sent))
    l = len(sent)
    f = df_sent.iloc[index_num].frequency
    unique_code = f"G{g}-L{l}-F{f}-{index_num}"
    sent_code_list.append([sent, f, unique_code])
   
if __name__ == '__main__':
    # with Pool(16) as p:
    with Pool(nprocs) as p: # Pool number CPU sayısına eşit olursa tüm CPU lar çalışır
        p.starmap(sentence_code, zip_list) # zip_list fonksiyon için çoklu eleman gerekli olduğu için kullanıldı.
# Not: starmap çoklu iterable değerlere ihtiyaç duyan fonksiyonlar için kullanılıyor ve bu değerler zip veya itertools şeklinde olmalı. O yüzden burada pool.map yerine pool.starmap kullanıldı. 

df_code = pd.DataFrame(list(sent_code_list), columns=["sentence", "frequency", "code"])
df_code.sort_values(by="frequency", ascending=False, inplace=True)
df_code

df_code.to_csv("Sentence_Merge_Code.csv", index=False)
